+++ src/App.js
@@ -1,9 +1,12 @@
+import { useDispatch } from 'react-redux';
+import { useEffect } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import './App.css';
 import { ToastContainer } from 'react-toastify';
 import Home from './routes/Home';
 import AppointmentList from './routes/AppointmentList';
-import DocDetails from './routes/DocDetails';
+import { doctorsFetchData } from './redux/doctors/doctorSlice';
+import DoctorDetails from './routes/DoctorDetails';
 import AppointmentDetails from './routes/AppointmentDetails';
 import SignUp from './routes/SignUp';
 import Login from './routes/Login';
@@ -11,6 +14,12 @@ import ProtectedRoute from './components/ProtectedRoute';
 import Doctors from './routes/Doctors';
 
 function App() {
+  const doctorsDispatch = useDispatch();
+  
+  useEffect(() => {
+    doctorsDispatch(doctorsFetchData());
+  }, [doctorsDispatch]);
+
   return (
     <>
       <BrowserRouter>
@@ -21,7 +30,7 @@ function App() {
           <Route element={<ProtectedRoute />}>
             <Route path="doctors" element={<Doctors />} />
             <Route path="appointment-list" element={<AppointmentList />} />
-            <Route path="docdetails" element={<DocDetails />} />
+            <Route path="/doctor/:id" element={<DoctorDetails />} />
             <Route path="appointment" element={<AppointmentDetails />} />
           </Route>
         </Routes>
